Error reponses in schema

+++ endpoint/printerbox.py
@@ -1,6 +1,6 @@
 import os
 from typing import List
-from fastapi import APIRouter, WebSocket, WebSocketDisconnect
+from fastapi import APIRouter, WebSocket, WebSocketDisconnect, status, HTTPException
 from pydantic import BaseModel
 from werkzeug.utils import secure_filename
 
@@ -10,10 +10,10 @@ from printer_code import PrinterCode
 
 
 class WSConnection():
-    websocket : WebSocket
-    printerCode : PrinterCode
+    websocket: WebSocket
+    printerCode: PrinterCode
 
-    def __init__(self, webSocket : WebSocket, printerCode : PrinterCode):
+    def __init__(self, webSocket: WebSocket, printerCode: PrinterCode):
         self.websocket = webSocket
         self.printerCode = printerCode
 
@@ -22,14 +22,14 @@ class WSConnectionManager:
     def __init__(self):
         self.connections: List[WSConnection] = []
 
-    async def connect(self, connection : WSConnection):
+    async def connect(self, connection: WSConnection):
         await connection.websocket.accept()
         self.connections.append(connection)
 
     def disconnect(self, websocket: WebSocket):
         self.connections.remove(websocket)
 
-    async def sendToPrinter(self, printerCode : PrinterCode, message: str):
+    async def sendToPrinter(self, printerCode: PrinterCode, message: str):
         for connection in self.connections:
             if connection.printerCode == printerCode:
                 await connection.websocket.send_text(message)
@@ -39,15 +39,14 @@ class WSConnectionManager:
             await connection.websocket.send_text(message)
 
 
-def allFilesInPrinterQueue(printerCode : PrinterCode):
-    files : List[FilePath] = []
-    
+def allFilesInPrinterQueue(printerCode: PrinterCode):
+    files: List[FilePath] = []
+
     for root, dirs, foundFiles in os.walk(queuesPath + printerCode, topdown=False):
         for name in foundFiles:
             files.append(FilePath(filename=os.path.join(root, name)))
-    
-    return files
 
+    return files
 
 
 router = APIRouter()
@@ -55,19 +54,23 @@ wsConnectionManager = WSConnectionManager()
 
 
 @router.get('/{printer_code}', response_model=List[FilePath])
-def printerboxQueue(printer_code : PrinterCode):
+def printerboxQueue(printer_code: PrinterCode):
     return allFilesInPrinterQueue(printer_code)
 
-@router.delete('/{printer_code}/{filename}', response_model=FilePath)
-def deleteNameTag(printer_code : PrinterCode, filename : str):
-    securedFileName = secure_filename(filename)
-    nameTagFilename = queuesPath + printer_code + '/' + os.path.basename(securedFileName)
-    
-    os.path.exists(nameTagFilename)
-    if os.path.isfile(nameTagFilename):
-        os.remove(nameTagFilename)
 
-    return FilePath(nameTagFilename)
+@router.delete('/{printer_code}/{filename}', response_model=FilePath)
+def deleteNameTag(printer_code: PrinterCode, filename: str):
+    securedFileName = secure_filename(filename)
+    nameTagFilename = queuesPath + printer_code + \
+        '/' + os.path.basename(securedFileName)
+
+    if os.path.exists(nameTagFilename) and os.path.isfile(nameTagFilename):
+        os.remove(nameTagFilename)
+        return FilePath(nameTagFilename)
+
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                        detail="NameTag not found")
+
 
 @router.websocket("/ws/{printer_code}")
 async def websocket_endpoint(websocket: WebSocket, printer_code: PrinterCode):
@@ -80,7 +83,7 @@ async def websocket_endpoint(websocket: WebSocket, printer_code: PrinterCode):
     try:
         while True:
             # Just reply what is received
-            data =await websocket.receive_text()
+            data = await websocket.receive_text()
             await wsConnectionManager.sendToPrinter(connection.printerCode, data)
     except WebSocketDisconnect:
         wsConnectionManager.disconnect(websocket)
